fix(test): move undici invariant to constructor, extract for safe test import (t/2113)
Two CI failures in PR #394:
 1. test-electron: static `import { assertUndiciMajorInvariant } from githubRestClient`
    triggered the module-level check at Vitest collection time → 0 tests collected.
 2. test-container: userland undici 8.9.0 vs node 22.23.2 bundled 6.28.0 → server crash
    on startup.

Fix:
 - Extract assertUndiciMajorInvariant to undiciInvariant.ts (pure fn, no side effects).
   Tests import from there; githubRestClient.ts re-exports for backward compat.
 - Move the runtime invariant check from module-level into GitHubRestClient constructor
   so importing the module (directly or transitively) never fires the check.
 - Pin undici to 6.28.0 in package.json to match node 22.23.2 bundled version so
   the constructor check passes in CI (node 22) and the container.
