+++ taxonomy-editor/package.json
@@ -67,7 +67,7 @@
     "react-dom": "^19.2.8",
     "react-markdown": "^10.1.0",
     "remark-gfm": "^4.0.1",
-    "undici": "8.9.0",
+    "undici": "6.28.0",
     "ws": "^8.21.1",
     "zod": "^4.4.3",
     "zustand": "^5.0.14"
